requirements.txt
streamlit>=1.32.0
gspread>=6.0.0
google-auth>=2.27.0
plotly>=5.18.0
numpy>=1.24.0
